Updated: MQL4/Signal Service Buddy v1_0.mq4 MQL4/Signal Service Buddy v1_0.ex4 README.md
Signal Service Buddy v1_0.mq4/ex4 - added the folder 'SQLite/' to the path
on line #51 as I had forgotten to add that. Without it, and using my
instructions in README.md, would result in an error in MetaTrader about
not finding the database file.

README.md - added a new Troubleshooting section, and a couple small
miscellaneous edits.